GH-74 Unsupported Type as a return type or field is not recognized
Moved tests into IT
Bumped commons-lang3 to prevent java11 failure
Bumped maven invoker plugin to fix issue where IT builds were not executed on java13
