+++ animal-sniffer-maven-plugin/pom.xml
@@ -77,7 +77,7 @@
     <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
-      <version>3.7</version>
+      <version>3.9</version>
     </dependency>
     <dependency>
       <groupId>org.apache.maven.shared</groupId>
