+++ animal-sniffer-maven-plugin/src/it/github-74-1/invoker.properties
@@ -1,0 +1,1 @@
+invoker.buildResult=failure
