+++ animal-sniffer-maven-plugin/src/it/github-74-1/src/main/java/localhost/IllegalFieldWithAccessorsSample.java
@@ -1,22 +1,29 @@
-package org.codehaus.mojo.animal_sniffer.samples;
+package localhost;
 
 import java.nio.file.Path;
 
 /**
  * @author Lukas Zaruba, lukas.zaruba@lundegaard.eu, 2019
  */
-public class IllegalFieldWithManipulationSample {
+public class IllegalFieldWithAccessorsSample {
 
     private String stringField;
     private Path pathField;
 
+    public String getStringField() {
+        return stringField;
+    }
+
+    public void setStringField(String stringField) {
+        this.stringField = stringField;
+    }
 
     public Path getPathField() {
         return pathField;
     }
 
     public void setPathField(Path pathField) {
-        this.pathField = pathField.resolve("other");
+        this.pathField = pathField;
     }
 
 }
