+++ animal-sniffer-maven-plugin/src/it/github-74-1/src/main/java/localhost/IllegalFieldWithManipulationSample.java
@@ -1,0 +1,21 @@
+package localhost;
+
+import java.nio.file.Path;
+
+/**
+ * @author Lukas Zaruba, lukas.zaruba@lundegaard.eu, 2019
+ */
+public class IllegalFieldWithManipulationSample {
+
+    private String stringField;
+    private Path pathField;
+
+    public Path getPathField() {
+        return pathField;
+    }
+
+    public void setPathField(Path pathField) {
+        this.pathField = pathField.resolve("other");
+    }
+
+}
