+++ animal-sniffer-maven-plugin/src/it/github-74-1/src/main/java/localhost/IllegalTypeReturn.java
@@ -1,4 +1,4 @@
-package org.codehaus.mojo.animal_sniffer.samples;
+package localhost;
 
 import java.nio.file.Path;
