+++ animal-sniffer-maven-plugin/src/it/github-74-2/src/main/java/localhost/FieldsOK.java
