+++ pom.xml
@@ -176,17 +176,19 @@
         <plugin>
           <groupId>org.apache.maven.plugins</groupId>
           <artifactId>maven-invoker-plugin</artifactId>
-          <version>3.0.1</version>
+          <version>3.2.1</version>
           <configuration>
             <projectsDirectory>src/it</projectsDirectory>
             <cloneProjectsTo>${project.build.directory}/it</cloneProjectsTo>
             <localRepositoryPath>${project.build.directory}/local-repo</localRepositoryPath>
             <settingsFile>src/it/settings.xml</settingsFile>
             <showErrors>true</showErrors>
-            <streamLogs>${invoker.streamLogs}</streamLogs>
+            <streamLogs>true</streamLogs>
             <postBuildHookScript>verify</postBuildHookScript>
-            <streamLogs>${invoker.streamLogs}</streamLogs>
-          </configuration>          
+            <goals>
+              <goal>verify</goal>
+            </goals>
+          </configuration>
         </plugin>
         <plugin>
           <groupId>org.codehaus.mojo</groupId>
